Add files via upload
Completed back end statistics
Larger issue discovered, if I write to the google sheets API more than 500 times in a minute and a half the API kicks me out.
Each time I receive a message the bot does about 50 writes so any faster than 1 message every 10 seconds and a crash occurs.
I have replicated this crash by sending a message every 6-9 seconds 10 times in a row.
I could probably solve this by updating the spreadsheet only when changes to specific cells occur, removing repetitive statistics, and putting more computational load on the program itself. For example only calculating the Top 5 counters when requested by command and not storing it in the sheet as this information is easy to recalculate.